Subscription confirmation support
Currently Zaqar can post notifications to the subscriber
once the subscription is created.
It doesn't need the subscriber to confirm the subscription.

This will lead a problem that users could send junk
information that whether subscriber want or not.

So we should support to confirm subscription by using
HTTP request like what Amazon SNS is doing.

Change-Id: I43c202e86a7a6a9605747096f707104d2d699010
Implements: blueprint subscription-confirmation-support