Spec of delayed queues support
Now one of the big function gaps of Zaqar is the delayed queue.
Currently, all the message posted to the queue will be visible
immediately. That's enough for most of the user cases. However,
for some user case, user want the message to be unavailable to
end users for a specific period of time.

Change-Id: I2610a8f89568f8af1cb292f0c8dea8f7deb4e4a3
blueprint: delayed-queues