Remove pool group from pool and flavor resource
Currently pool group is used in pool and flavor resource,
but the pool group only supports a 1:1 mapping with flavor.
So it's not necessary to keep it since we can map
1 flavor : N pool directly.

For making a clarification to end user,
this bp proposes to remove useless pool group from Zaqar,
just keeps the pool resource.

Change-Id: I5927aea4e1ffa632bb78e3fa9d340c3d812bad2a
Implements: blueprint remove-pool-group-from-zaqar